// src/app/core/enums/navigation-links.ts
export enum NavigationLink {
  Home = '',
  Hero = 'hero',
  List = 'list',
  NewForm = 'form',
  EditForm = 'editForm',
}

export enum PathParams {
  HeroId = 'heroId',
}

export enum TitlesPage {
  Home = 'Gestion de heroes',
  HerosList = 'Listado de heroes',
  HeroInsert = 'Creación de nuevo heroe',
  HeroUpdate = 'Actualizacion de heroe',
}